requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.3
